1. fix css 2. add press page
