+++ src/components/Footer/index.js
@@ -117,127 +117,129 @@ const Page = ({ children, ...props }) => (
   </Link>
 )
 
-const Footer = () => {
+const UpperFooter = () => (
+  <div css={upperFooterContainer}>
+    <div css={contentContainer}>
+      <div css={contentLeftContainer}>
+        <div>
+          <img css={maiImage} src={mailogo} alt="MAI" />
+          <div css={poweredByContainer}>
+            <p css={theme => [fontStyle(theme), poweredBy]}>
+              <FormattedMessage id="poweredBy" defaultMessage="Powered by" />
+            </p>
+            <img css={mdtImage} src={mdtLogo} alt="MDT" />
+          </div>
+        </div>
+        <div css={logosContainer}>
+          <SocialIcon src={linkedinIcon} alt="linkedin" />
+          <SocialIcon src={mediumIcon} alt="medium" />
+          <SocialIcon src={twitterIcon} alt="twitter" />
+          <SocialIcon src={facebookIcon} alt="facebook" />
+          <SocialIcon src={wechatIcon} alt="wechat" />
+          <SocialIcon src={weiboIcon} alt="weibo" last={true} />
+        </div>
+      </div>
+      <div css={contentRightContainer}>
+        <Page to="/jobs">
+          <OtherPages>
+            <FormattedMessage id="jobs" defaultMessage="JOBS" />
+          </OtherPages>
+        </Page>
+        <Page to="/press">
+          <OtherPages>
+            <FormattedMessage id="press" defaultMessage="PRESS" />
+          </OtherPages>
+        </Page>
+        <Page to="/privacyPolicy">
+          <OtherPages>
+            <FormattedMessage
+              id="privacyPolicy"
+              defaultMessage="PRIVACY  POLICY"
+            />
+          </OtherPages>
+        </Page>
+        <Page to="/termsofuse">
+          <OtherPages>
+            <FormattedMessage id="termsOfUse" defaultMessage="TERMS OF USE" />
+          </OtherPages>
+        </Page>
+      </div>
+    </div>
+    <p css={theme => [fontStyle(theme), allRightsReserved]}>
+      <FormattedMessage
+        id="AllRightsReserved"
+        defaultMessage="ALL RIGHTS RESERVED © 2019 Measurable AI"
+      />
+    </p>
+  </div>
+)
+
+const LowerFooter = () => {
   const [state, dispatch] = useReducer(reducer, initialState)
   const onClose = useCallback(() => {
     dispatch({ type: CLOSE })
   }, [])
-
   return (
-    <footer>
-      <div css={upperFooterContainer}>
-        <div css={contentContainer}>
-          <div css={contentLeftContainer}>
-            <div>
-              <img css={maiImage} src={mailogo} alt="MAI" />
-              <div css={poweredByContainer}>
-                <p css={theme => [fontStyle(theme), poweredBy]}>
-                  <FormattedMessage
-                    id="poweredBy"
-                    defaultMessage="Powered by"
-                  />
-                </p>
-                <img css={mdtImage} src={mdtLogo} alt="MDT" />
-              </div>
-            </div>
-            <div css={logosContainer}>
-              <SocialIcon src={linkedinIcon} alt="linkedin" />
-              <SocialIcon src={mediumIcon} alt="medium" />
-              <SocialIcon src={twitterIcon} alt="twitter" />
-              <SocialIcon src={facebookIcon} alt="facebook" />
-              <SocialIcon src={wechatIcon} alt="wechat" />
-              <SocialIcon src={weiboIcon} alt="weibo" last={true} />
-            </div>
-          </div>
-          <div css={contentRightContainer}>
-            <Page to="/jobs">
-              <OtherPages>
-                <FormattedMessage id="jobs" defaultMessage="JOBS" />
-              </OtherPages>
-            </Page>
-            <Page to="/press">
-              <OtherPages>
-                <FormattedMessage id="press" defaultMessage="PRESS" />
-              </OtherPages>
-            </Page>
-            <Page to="/privacyPolicy">
-              <OtherPages>
-                <FormattedMessage
-                  id="privacyPolicy"
-                  defaultMessage="PRIVACY  POLICY"
-                />
-              </OtherPages>
-            </Page>
-            <Page to="/termsofuse">
-              <OtherPages>
-                <FormattedMessage
-                  id="termsOfUse"
-                  defaultMessage="TERMS OF USE"
-                />
-              </OtherPages>
-            </Page>
-          </div>
-        </div>
-        <p css={theme => [fontStyle(theme), allRightsReserved]}>
-          <FormattedMessage
-            id="AllRightsReserved"
-            defaultMessage="ALL RIGHTS RESERVED © 2019 Measurable AI"
-          />
-        </p>
-      </div>
-      <div css={container}>
-        <div css={content}>
-          <Popup
-            key={state.closeCount}
-            trigger={
-              <button css={helpButton}>
-                <span css={helpSymbol}>?</span>HELP
+    <div css={container}>
+      <div css={content}>
+        <Popup
+          key={state.closeCount}
+          trigger={
+            <button css={helpButton}>
+              <span css={helpSymbol}>?</span>
+              <FormattedMessage id="help" defaultMessage="HELP" />
+            </button>
+          }
+          closeOnDocumentClick
+          contentStyle={popoverContainer}
+          position="top center"
+          onClose={onClose}
+        >
+          <>
+            <div css={popoverHeader}>
+              <FormattedMessage id="help" defaultMessage="HELP" />
+              <button css={closeButton} onClick={onClose}>
+                <div css={closeSymbol} />
               </button>
-            }
-            closeOnDocumentClick
-            contentStyle={popoverContainer}
-            position="top center"
-            onClose={onClose}
-          >
-            <>
-              <div css={popoverHeader}>
-                HELP
-                <button css={closeButton} onClick={onClose}>
-                  <div css={closeSymbol} />
-                </button>
-              </div>
-              <div css={popoverBody}>
-                {state.page === PAGE_CALLBACK && <Callback />}
-                {!state.page && (
-                  <>
-                    <MenuItem
-                      onClick={() => dispatch({ type: GO_TO_CALLBACK })}
-                      src={CallbackIconSrc}
-                      text={
-                        <FormattedMessage
-                          id="requestCallback"
-                          defaultMessage="Request a callback"
-                        />
-                      }
-                    />
-                    <MenuItem
-                      src={MessageIconSrc}
-                      text={
-                        <FormattedMessage
-                          id="leaveMessage"
-                          defaultMessage="Leave a message"
-                        />
-                      }
-                    />
-                  </>
-                )}
-              </div>
-            </>
-          </Popup>
-        </div>
+            </div>
+            <div css={popoverBody}>
+              {state.page === PAGE_CALLBACK && <Callback />}
+              {!state.page && (
+                <>
+                  <MenuItem
+                    onClick={() => dispatch({ type: GO_TO_CALLBACK })}
+                    src={CallbackIconSrc}
+                    text={
+                      <FormattedMessage
+                        id="requestCallback"
+                        defaultMessage="Request a callback"
+                      />
+                    }
+                  />
+                  <MenuItem
+                    src={MessageIconSrc}
+                    text={
+                      <FormattedMessage
+                        id="leaveMessage"
+                        defaultMessage="Leave a message"
+                      />
+                    }
+                  />
+                </>
+              )}
+            </div>
+          </>
+        </Popup>
       </div>
-    </footer>
+    </div>
   )
 }
+
+const Footer = () => (
+  <footer>
+    <UpperFooter />
+    <LowerFooter />
+  </footer>
+)
 
 export default Footer
