+++ src/intl/zh-HK.json
@@ -73,6 +73,7 @@
   },
   "formSubmitSuccess": "Your request has been sent!",
   "formSubmitFailed": "We failed to send your request. Please try again later!",
+  "help": "HELP",
 
   "weAreHiring": "We’re HIRING!",
   "joinUs": "Join us in our quest to make communication friendly and efficient! We're looking for full-stack developers to our team in Hong Kong, and we're offering generous equity alongside competitive pay. Not to mention free meals, snacks and drinks, gym room and table tennis, and happy hours!",
